Challenge Task: Simulated plants
The location of the plants are kept in a HashMap in the field class.

The lamb and cow class now have a findFood method just like the predator class.

There is a plant in every cell, the plants will take at least two steps to grow in order for it to be edible by animals. Whenever a plant is eaten, the plant growth will be reseted back to 0.